fix random mat type bug

+++ bandwidth/kernel.cl
@@ -5,16 +5,16 @@ __kernel void global_bandwidth_vec1(const int heightA, const int widthA, __globa
     b[idx] = a[idx];
 }
 
-
-__kernel void global_bandwidth_vec2(const int heightA, const int widthA, __global const float *a, __global float *b) {
+/*
+__kernel void global_bandwidth_vec2(const int heightA, const int widthA, __global const half *a, __global half *b) {
     const int idx = get_global_id(0);
     const int step = idx << 1;
 
     CL_ELEM_TYPE value = *((__global CL_ELEM_TYPE *)(a + step));
     *((__global CL_ELEM_TYPE *)(b + step)) = value;
 }
+*/
 
-/*
 __kernel void global_bandwidth_vec2(const int heightA, const int widthA, __global const CL_ELEM_TYPE *a, __global CL_ELEM_TYPE *b) {
     const int idx = get_global_id(0);
     const int step = idx;// << 1;
@@ -22,8 +22,6 @@ __kernel void global_bandwidth_vec2(const int heightA, const int widthA, __globa
     CL_ELEM_TYPE value = *((__global CL_ELEM_TYPE *)(a + step));
     *((__global CL_ELEM_TYPE *)(b + step)) = value;
 }
-*/
-
 
 __kernel void global_bandwidth_vec4(const int heightA, const int widthA, __global const CL_ELEM_TYPE *a, __global CL_ELEM_TYPE *b) {
     const int idx = get_global_id(0);
